Move DataStoreEvent into persistence (for now)
